DOC: Use the DWI-computed braimask for UKF tractography
Use the DWI-computed brainmask for UKF tractography: the QSIPrep version
used when the pipeline was initially developed (0.19.0) worked
correctly employing the T1w-derived brainmask for UKF tractography, but
in the version that is being used currently (1.0.0rc1) the size of that
mask does not match the DWI data/is not aligned with the DWI data, and
thus UKF provides unexpected results.

As mostly default flags have been used to preprocess the data, the DWI
data are only aligned into ACPC, and not to the T1w data.